refactor: split quadlet/mod.rs and compose/mod.rs under the 500-line limit (#358)
Two module roots had grown past the 500-line hard limit. Pure
relocation, no behavior or public-API change; all existing tests keep
passing.

- `internal/quadlet/mod.rs` (636 → 106): inline tests moved into a
`tests/` submodule, split by theme into `units.rs` and `fields.rs` with
the shared lookup helper in `tests/mod.rs`.
- `internal/compose/mod.rs` (561 → 219):
`resolve_order`/`resolve_levels` extracted to `order.rs`, the YAML
alias-expansion guard and merge-key application to `merge.rs`.
`compose::resolve_order` and the crate-root re-exports are unchanged.

No file in either tree now exceeds the limit (largest new file:
`quadlet/tests/units.rs` at 268).

Closes #355

---------
