+++ internal/quadlet/tests/mod.rs
@@ -1,0 +1,11 @@
+use crate::quadlet::{QuadletOutput, QuadletUnit};
+
+mod fields;
+mod units;
+
+fn unit_named<'a>(out: &'a QuadletOutput, filename: &str) -> &'a QuadletUnit {
+	out.units
+		.iter()
+		.find(|u| u.filename == filename)
+		.unwrap_or_else(|| panic!("no unit named {filename}"))
+}
